requirements.txt
Django==4.2
Pillow==10.1.0
opencv-python-headless==4.8.1.78
opencv-contrib-python-headless==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
dlib
cmake
gunicorn
whitenoise==6.6.0
python-dotenv==1.0.0